+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.21.0](https://github.com/djm204/agentic-team-templates/compare/agentic-team-templates-v0.20.0...agentic-team-templates-v0.21.0) (2026-02-01)
+
+
+### Features
+
+* **templates:** add 16 new agentic archetype templates ([#69](https://github.com/djm204/agentic-team-templates/issues/69)) ([6587ce2](https://github.com/djm204/agentic-team-templates/commit/6587ce2d6ec2afcc3b9a7fb19e13e60c0c8f608c))
+
+## [0.20.0](https://github.com/djm204/agentic-team-templates/compare/agentic-team-templates-v0.19.1...agentic-team-templates-v0.20.0) (2026-02-01)
+
+
+### Features
+
+* **templates:** add project-manager, supply-chain, executive-assistant, grant-writer templates ([#66](https://github.com/djm204/agentic-team-templates/issues/66)) ([aae6243](https://github.com/djm204/agentic-team-templates/commit/aae62431c5f677adce3ba0d009f9ced89ba0728f))
+
+## [0.19.1](https://github.com/djm204/agentic-team-templates/compare/agentic-team-templates-v0.19.0...agentic-team-templates-v0.19.1) (2026-01-31)
+
+
+### Miscellaneous
+
+* update CLAUDE.md and release-please.yml ([#64](https://github.com/djm204/agentic-team-templates/issues/64)) ([48a488e](https://github.com/djm204/agentic-team-templates/commit/48a488e2812672a4c36a4e7744eef9c4da16e285))
+
 ## [0.19.0](https://github.com/djm204/agentic-team-templates/compare/agentic-team-templates-v0.18.0...agentic-team-templates-v0.19.0) (2026-01-31)